Restructure eigen installation using upstream as a base
This keeps only what was missing (at least in my version
of eigen3 -- ubuntu 18.04) from the upstream library and
reorganises the installation process to install the
libeigen.a artifact among the rest of the eigen package
without obscure copies. I have tested this with owl and
seems to be working.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>