New attempt, using C++ all the way
Note that there are going to be a number of warnings of the type
warning: empty struct has size 0 in C, size 1 in C++
This is beacuse we are using an extension from gcc for the empty structs, which are otherwise not allowed by C syntax

Maybe we should reconsider them.

Also, the hack on the stubs will require bash and sed, so will likely be a problem for a native (not cygwin) windows version.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>